feat(webpack-builder): isolate `stub` module from other modules in `src`
