+++ packages/builder/webpack-builder/src/shared/constants.ts
@@ -36,13 +36,3 @@ export const MEDIA_EXTENSIONS = [
 ];
 
 export const RUNTIME_CHUNK_NAME = 'builder-runtime';
-
-export const STUB_BUILDER_PLUGIN_BUILTIN =
-  (
-    {
-      true: true,
-      false: false,
-      default: 'default',
-      minimal: 'minimal',
-    } as const
-  )[process.env.STUB_BUILDER_PLUGIN_BUILTIN || ''] || undefined;
