+++ packages/builder/webpack-builder/src/stub/constants.ts
@@ -1,0 +1,9 @@
+export const STUB_BUILDER_PLUGIN_BUILTIN =
+  (
+    {
+      true: true,
+      false: false,
+      default: 'default',
+      minimal: 'minimal',
+    } as const
+  )[process.env.STUB_BUILDER_PLUGIN_BUILTIN || ''] || undefined;
